# api/openapi.yml
openapi: 3.0.3
info:
  title: Fansworld Creator Platform API
  version: 1.0.0
  description: |
    Production-ready API for creator subscription platform with OnlyFans-level functionality.
    Supports content creation, subscription management, messaging, and creator monetization.
  contact:
    name: Fansworld API Team
    email: api@fansworld.com
  license:
    name: Proprietary
    url: https://fansworld.com/license
servers:
  - url: https://api.fansworld.com/v1
    description: Production server
  - url: https://staging-api.fansworld.com/v1
    description: Staging server
  - url: http://localhost:3001/api/v1
    description: Development server
security:
  - bearerAuth: []
paths:
...
  /media/sign-upload:
    post:
      summary: Get signed URL for media upload
      tags: [Media]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [filename, contentType, size]
              properties:
                filename: { type: string }
                contentType: { type: string }
                size: { type: integer, maximum: 104857600 } # 100MB
                isPublic: { type: boolean, default: false }
      responses:
        '201':
          description: Signed upload URL generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignedUploadResponse'
  /media/{mediaId}:
    get:
      summary: Get media file (with access control)
      tags: [Media]
      parameters:
        - in: path
          name: mediaId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Media file
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '402':
          description: Payment required
        '404':
          description: Media not found

  # Subscriptions (Stubbed)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SignedUploadResponse:
      type: object
      properties:
        uploadUrl: { type: string, format: uri }
        mediaId: { type: string }
        fields: { type: object }
        expiresIn: { type: integer }

    # Subscription Schemas
...